+++ packages/seedless-onboarding-controller/src/SeedlessOnboardingController.test.ts
@@ -1083,9 +1083,7 @@ describe('SeedlessOnboardingController', () => {
           }),
         },
         async ({ controller }) => {
-          expect(
-            await controller.checkIsSeedlessOnboardingUserAuthenticated(),
-          ).toBe(true);
+          expect(await controller.getIsUserAuthenticated()).toBe(true);
         },
       );
     });
@@ -1093,18 +1091,13 @@ describe('SeedlessOnboardingController', () => {
     it('should return false if the user is not authenticated (accessToken is missing)', async () => {
       await withController(
         {
-          state: {
-            userId,
-            authConnectionId,
-            groupedAuthConnectionId,
-            metadataAccessToken,
-            revokeToken,
-          },
+          state: getMockInitialControllerState({
+            withMockAuthenticatedUser: true,
+            withoutMockAccessToken: true, // missing accessToken
+          }),
         },
         async ({ controller }) => {
-          expect(
-            await controller.checkIsSeedlessOnboardingUserAuthenticated(),
-          ).toBe(false);
+          expect(await controller.getIsUserAuthenticated()).toBe(false);
         },
       );
     });
@@ -1112,28 +1105,57 @@ describe('SeedlessOnboardingController', () => {
     it('should return false if the user is not authenticated (revokeToken is missing)', async () => {
       await withController(
         {
-          state: {
-            userId,
-            authConnectionId,
-            groupedAuthConnectionId,
-            metadataAccessToken,
-            accessToken,
-          },
+          state: getMockInitialControllerState({
+            withMockAuthenticatedUser: true,
+            withoutMockRevokeToken: true, // missing revokeToken
+          }),
         },
         async ({ controller }) => {
-          expect(
-            await controller.checkIsSeedlessOnboardingUserAuthenticated(),
-          ).toBe(false);
+          expect(await controller.getIsUserAuthenticated()).toBe(false);
         },
       );
     });
 
     it('should return false if the user is not authenticated (social login details are missing)', async () => {
       await withController(async ({ controller }) => {
-        expect(
-          await controller.checkIsSeedlessOnboardingUserAuthenticated(),
-        ).toBe(false);
+        expect(await controller.getIsUserAuthenticated()).toBe(false);
       });
+    });
+
+    it('should skip the `accessToken` and `revokeToken` check if `skipVaultCreationCheck` is true', async () => {
+      await withController(
+        {
+          state: getMockInitialControllerState({
+            withMockAuthenticatedUser: true,
+            withoutMockAccessToken: true, // missing accessToken
+            withoutMockRevokeToken: true, // missing revokeToken
+          }),
+        },
+        async ({ controller }) => {
+          const skipVaultCreationCheck = true;
+          expect(
+            await controller.getIsUserAuthenticated(skipVaultCreationCheck),
+          ).toBe(true);
+        },
+      );
+    });
+
+    it('should return false if the user is not authenticated with `skipVaultCreationCheck` is true.', async () => {
+      await withController(
+        {
+          state: {
+            // missing NodeAuthTokens, authConnectionId, userId
+            groupedAuthConnectionId,
+            metadataAccessToken,
+          },
+        },
+        async ({ controller }) => {
+          const skipVaultCreationCheck = true;
+          expect(
+            await controller.getIsUserAuthenticated(skipVaultCreationCheck),
+          ).toBe(false);
+        },
+      );
     });
   });